requirements.txt
fastapi==0.101.1
uvicorn==0.28.1
requests
httpx
python-dotenv
python-multipart